Fix number formatting on non-64bit architectures (#235)
Fix number formatting on non-64bit architectures (#234)

Run CI tests with GOARCH=386 (-race is not supported when
running tests on Linux/386.)

---------

+++ interp/vm.go
@@ -980,7 +980,7 @@ func (p *interp) callBuiltin(builtinOp compiler.BuiltinOp) error {
 		p.replaceTop(num(float64(index + 1)))
 
 	case compiler.BuiltinInt:
-		p.replaceTop(num(float64(int(p.peekTop().num()))))
+		p.replaceTop(num(float64(int64(p.peekTop().num()))))
 
 	case compiler.BuiltinLength:
 		p.push(num(float64(len(p.line))))
